Fix: [all] fix compatibility problems for low-end devices

+++ include/ifrit/core/math/LowDiscrepancy.h
@@ -34,7 +34,7 @@ namespace Ifrit::Math
         bits = ((bits & 0x33333333u) << 2u) | ((bits & 0xCCCCCCCCu) >> 2u);
         bits = ((bits & 0x0F0F0F0Fu) << 4u) | ((bits & 0xF0F0F0F0u) >> 4u);
         bits = ((bits & 0x00FF00FFu) << 8u) | ((bits & 0xFF00FF00u) >> 8u);
-        return f32(bits) * 2.3283064365386963e-10; // / 0x100000000
+        return static_cast<f32>(f64(bits) * 2.3283064365386963e-10); // / 0x100000000
     }
 
     IF_FORCEINLINE Vector2f Hammersley2d(u32 x, u32 N)
